Don't convert $Shape to Partial
TypeScript's Partial isn't the same as Flow's $Shape. The latter is an
unsafe escape hatch that accepts missing properties but allows them to
be treated as non-missing where the property is accessed. TypeScript
doesn't have an equivalent feature, so any sensible output we use will
either generate some errors or be overly lax and allow completely
incorrect code.

This patch changes the conversion so that $Shape<T> is converted to T,
instead of the problematic Partial<T>. This means that now construction
sites of values that previously constructed values of $Shape<T> will
give errors, instead of the use sites of old $Shape<T> values like with
the Partial<T> version.

Based on our little experiments, this behaviour results in less errors
in our existing codebase in practice, and those error cases can be
worked around with `as` upcasts.